# postman/specs/Legal API/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Legal API",
    "version": "1.0.0",
    "description": "API documentation for Legal backend"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ObjectId": {
        "type": "string",
        "example": "681d0f8f8d7c4b0012f34abc"
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 100
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "example": 10
          }
        }
      },
...
      "PaymentType": {
        "type": "string",
        "enum": [
          "income",
          "expense"
        ],
        "example": "income"
      },
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "pending",
          "completed",
          "cancelled"
        ],
        "example": "completed"
      },
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "cash",
          "transfer",
          "card",
          "check",
          "other"
        ],
        "example": "transfer"
      },
      "FinanceSourceType": {
        "type": "string",
        "enum": [
          "manual",
          "case_payment"
        ],
        "example": "manual"
      },
...
      "FinanceTransaction": {
        "type": "object",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "title": {
            "type": "string",
            "example": "Pago de honorarios"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "example": 1200
          },
          "type": {
            "$ref": "#/components/schemas/PaymentType"
          },
          "category": {
            "type": "string",
            "example": "Honorarios"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "transactionDate": {
            "type": "string",
            "format": "date-time"
          },
          "registeredBy": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "updatedBy": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "deletedBy": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "relatedCaseId": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "sourceType": {
            "$ref": "#/components/schemas/FinanceSourceType"
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateFinanceTransactionDTO": {
        "type": "object",
        "required": [
          "title",
          "amount",
          "type",
          "category"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "Pago de honorarios"
          },
          "description": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "type": {
            "$ref": "#/components/schemas/PaymentType"
          },
          "category": {
            "type": "string"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "reference": {
            "type": "string"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "transactionDate": {
            "type": "string",
            "format": "date-time"
          },
          "relatedCaseId": {
            "type": "string"
          }
        }
      },
      "UpdateFinanceTransactionDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "type": {
            "$ref": "#/components/schemas/PaymentType"
          },
          "category": {
            "type": "string"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "reference": {
            "type": "string"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "transactionDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FinanceTransactionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "transaction": {
            "$ref": "#/components/schemas/FinanceTransaction"
          }
        }
      },
      "FinanceTransactionListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FinanceTransaction"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/finance-transactions": {
      "post": {
        "summary": "Crear una transacción financiera",
        "tags": [
          "Finance Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFinanceTransactionDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transacción creada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceTransactionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "get": {
        "summary": "Obtener transacciones financieras",
        "tags": [
          "Finance Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por tipo"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por estado"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por categoría"
          },
          {
            "in": "query",
            "name": "relatedCaseId",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por caso relacionado"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "number"
            },
            "description": "Número de página"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "number"
            },
            "description": "Cantidad de resultados"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de transacciones obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceTransactionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/finance-transactions/{transactionId}": {
      "get": {
        "summary": "Obtener transacción por ID",
        "tags": [
          "Finance Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la transacción"
          }
        ],
        "responses": {
          "200": {
            "description": "Transacción obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceTransactionResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Transacción no encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualizar transacción financiera",
        "tags": [
          "Finance Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la transacción"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFinanceTransactionDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transacción actualizada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceTransactionResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Transacción no encontrada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar transacción financiera",
        "tags": [
          "Finance Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la transacción"
          }
        ],
        "responses": {
          "200": {
            "description": "Transacción eliminada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Transacción eliminada correctamente"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Transacción no encontrada"
          }
        }
      }
    },
...
  },
...
}